Move getSignedDiscoveryToken helper to helper_test.go
